refactor: prep for semantics crate split
