+++ crates/semantics/src/checker/sealing.rs
